refactor: add google auth response dto to swagger
